PR : release ver 1.0

+++ src/main/java/com/sparta/abnb/service/CommentService.java
@@ -4,9 +4,11 @@ import com.sparta.abnb.dto.requestdto.CommentRequestDto;
 import com.sparta.abnb.dto.responsedto.CommentDto;
 import com.sparta.abnb.dto.responsedto.CommentResponseDto;
 import com.sparta.abnb.entity.Comment;
+import com.sparta.abnb.entity.Reservation;
 import com.sparta.abnb.entity.Room;
 import com.sparta.abnb.entity.User;
 import com.sparta.abnb.repository.CommentRepository;
+import com.sparta.abnb.repository.ReservationRepository;
 import com.sparta.abnb.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -23,6 +25,8 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final RoomRepository roomRepository;
+    private final ReservationRepository reservationRepository;
+    private final RoomService roomService;
 
 
     //후기 작성
@@ -30,6 +34,13 @@ public class CommentService {
         // 해당 room이 존재하는지 확인
         Room room = roomRepository.findById(roomId).orElseThrow(()
                 -> new IllegalArgumentException("찾으시는 ROOM은 존재하지 않습니다."));
+
+        // 해당 user가 같은 Room을 여러번 예약할 수도 있을 경우를 생각
+        List<Reservation> reservations = reservationRepository.findAllByUserAndRoom(user,room);
+        if(reservations.isEmpty()){
+            throw new IllegalArgumentException("해당 ROOM 이용 기록이 존재하지 않습니다.");
+        }
+
 
         // 댓글 작성 및 저장
         Comment comment = Comment.builder()
@@ -43,6 +54,9 @@ public class CommentService {
         CommentDto commentResponseDto = CommentDto.builder()
                 .commentId(comment.getCommentId())
                 .comment(comment.getComment())
+                .profilePicture(comment.getUser().getProfilePicture())
+                .username(comment.getUser().getUsername())
+                .createdAt(comment.getCreatedAt())
                 .build();
 
         return ResponseEntity.status(HttpStatus.OK).body(commentResponseDto);
@@ -50,8 +64,9 @@ public class CommentService {
     }
 
     // 해당 Room의 전체 후기 조회
-    public CommentResponseDto getComments(Long roomId){
-        List<Comment> commentList = commentRepository.findByRoomId(roomId);
+    public ResponseEntity<CommentResponseDto> getComments(Long roomId){
+        Room room = roomService.findRoom(roomId);
+        List<Comment> commentList = commentRepository.findAllByRoom(room);
 
         if(commentList.isEmpty()){
             throw new IllegalArgumentException("해당 ROOM은 후기가 존재하지 않습니다.");
@@ -67,10 +82,12 @@ public class CommentService {
                 .collect(Collectors.toList());
 
         //commentList -> Dto
-        return CommentResponseDto.builder()
+        CommentResponseDto commentResponseDto = CommentResponseDto.builder()
                 .totalComments(commentList.size())
                 .commentResponseDtos(commentDtos)
                 .build();
+
+        return ResponseEntity.status(HttpStatus.OK).body(commentResponseDto);
     }
 
     //후기 삭제
@@ -84,6 +101,11 @@ public class CommentService {
         // 해당 후기를 DB에 있는지 확인
         Comment comment = commentRepository.findById(commentId).orElseThrow(()->
                 new IllegalArgumentException("작성하신 후기가 존재하지 않습니다."));
+
+        // 후기를 작성한 User인지 확인
+        if(!comment.getUser().getUserId().equals(user.getUserId())){
+            throw new IllegalArgumentException("후기를 작성한 USER가 아닙니다.");
+        }
 
         commentRepository.delete(comment);
